comment exception handle

+++ src/main/java/life/coder/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package life.coder.community.controller;
 
 import life.coder.community.dto.QuestionDTO;
-import life.coder.community.mapper.QuestionMapper;
 import life.coder.community.model.Question;
 import life.coder.community.model.User;
 import life.coder.community.service.QuestionService;
@@ -22,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                      Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -42,7 +41,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model){
